+++ src/th06_102h.exe/MainMenu/DrawReplayMenu.c
@@ -3,7 +3,7 @@ undefined4 __thiscall MainMenu::DrawReplayMenu(MainMenu *this)
 
 {
   int replay_amount;
-  int i;
+  uint i;
   AnmVm *vm_ref;
   
   AsciiManager::AddFormatText
@@ -11,7 +11,8 @@ undefined4 __thiscall MainMenu::DrawReplayMenu(MainMenu *this)
                     /* unsure */
   replay_amount = this->chosenReplay - this->chosenReplay % 15;
   vm_ref = &this->field98_0x6820;
-  for (i = replay_amount; (i < replay_amount + 0xf && (i < this->replayFilesNum)); i = i + 1) {
+  for (i = replay_amount; ((int)i < replay_amount + 0xf && ((int)i < this->replayFilesNum));
+      i = i + 1) {
     if ((g_Supervisor.cfg.opts >> 8 & 1 | g_Supervisor.cfg.opts & 1) == 0) {
       if (i == this->chosenReplay) {
         g_AsciiManager.color = 0xffff8080;
@@ -49,7 +50,7 @@ undefined4 __thiscall MainMenu::DrawReplayMenu(MainMenu *this)
                (double)*(float *)&this->currentReplay->field_0x2c);
     AsciiManager::AddFormatText(&g_AsciiManager,&(this->field114_0x7920).pos,"Stage  LastScore");
     vm_ref = &this->field114_0x7920;
-    for (i = 0; i < 7; i = i + 1) {
+    for (i = 0; (int)i < 7; i = i + 1) {
       if ((g_Supervisor.cfg.opts >> NO_COLOR_COMP & 1 | g_Supervisor.cfg.opts & 1) ==
           USE_D3D_HW_TEXTURE_BLENDING) {
         if (i == this->cursor) {
